+++ data/data.go
@@ -5,6 +5,7 @@ package data
 import "embed"
 
 //go:embed packagers/default.yaml
+//go:embed packagers/bcd-demo.yaml
 //go:embed tables/response-codes.yaml
 //go:embed tables/mcc.yaml
 //go:embed tables/processing-code.yaml